Change lastFMGetGenresByAlbum to catch API errors
If the Last.fm API can't find an album, the object it returns contains different properties: an error message & code number in place of expected results.

Change lastFMGetGenresByAlbum so it logs an error message before checking the tag objects for data (b/c if the API returns an error, the tags object won't exist, and the broswer will throw an error).